requirements.txt
asgiref==3.9.2
black==25.9.0
click==8.3.0
Django==5.2.7
django-environ==0.12.0
djangorestframework==3.15.2
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
django-filter==24.2
django-ratelimit==4.1.0
drf-spectacular==0.27.0
iniconfig==2.1.0
mypy==1.18.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
Pillow==11.0.0
platformdirs==4.4.0
pluggy==1.6.0
psycopg==3.2.10
psycopg-binary==3.2.10
Pygments==2.19.2
pytest==8.4.2
pytest-django==4.11.1
pytokens==0.1.10
ruff==0.13.3
sqlparse==0.5.3
typing_extensions==4.15.0
whitenoise==6.11.0
django-hugeicons-stroke==1.0.0
